Perfomance Improvements
Turns out having LoadModule perframe is not the best idea ever. It takes a lot of cycles to call it, load it, and then return the result.

Please do not repeat what I removed in this commit.

Also removed the option for players to join mid game since that wasn't on the original theme.